Move ATTACK_PREVIEW_BANNER to context
